fix: init mango indexes only for supported couchdb versions

+++ src/couch/init.js
@@ -108,8 +108,13 @@ async function checkSecurity(db, admin) {
 async function initDesignDocs(couch) {
   const db = couch._db;
   const { dbName } = db;
+  const [major, minor, patch] = await couch._nano.getVersion();
+  const supportsIndexes = major >= 3 || (major === 2 && minor >= 2);
+  debug.trace(
+    `couchdb version is ${major}.${minor}.${patch}, mango indexes are ${supportsIndexes ? '' : 'not '}supported`,
+  );
   debug.trace('check design documents for database %s', dbName);
-  var custom = couch._customDesign;
+  let custom = couch._customDesign;
   custom.views = custom.views || {};
   custom.indexes = custom.indexes || {};
   const designNames = new Set();
@@ -145,28 +150,33 @@ async function initDesignDocs(couch) {
   }
 
   // Create new indexes
-  for (let indexName of Object.keys(custom.indexes)) {
-    const index = custom.indexes[indexName];
-    index.name = indexName;
-    await db.createIndex(index);
-  }
-
-  // Delete orphan indexes
-  const allDesignDocs = await db.getDesignDocs();
-  for (let designDoc of allDesignDocs) {
-    if (designDoc.language === 'query') {
-      // if design doc is not referenced, delete the whole design document
-      if (!indexDesignNames.has(designDoc._id.replace(/^_design\//, ''))) {
-        debug.warn('destroy design doc', designDoc._id);
-        await db.destroyDocument(designDoc._id, designDoc._rev);
-      } else {
-        for (let key of Object.keys(designDoc.views)) {
-          if (!custom.indexes[key]) {
-            await db.deleteIndex(designDoc._id, key);
+  if (supportsIndexes) {
+    for (let indexName of indexNames) {
+      const index = custom.indexes[indexName];
+      index.name = indexName;
+      await db.createIndex(index);
+    }
+    // Delete orphan indexes and design docs
+    const allDesignDocs = await db.getDesignDocs();
+    for (let designDoc of allDesignDocs) {
+      if (designDoc.language === 'query') {
+        // if design doc is not referenced, delete the whole design document
+        if (!indexDesignNames.has(designDoc._id.replace(/^_design\//, ''))) {
+          debug.warn('destroy design doc', designDoc._id);
+          await db.destroyDocument(designDoc._id, designDoc._rev);
+        } else {
+          for (let key of Object.keys(designDoc.views)) {
+            if (!custom.indexes[key]) {
+              await db.deleteIndex(designDoc._id, key);
+            }
           }
         }
       }
     }
+  } else if (indexNames.length > 0) {
+    debug.error(
+      `The database ${couch._databaseName} has configured mango indexes but they are not supported by rest-on-couch with CouchDB version ${major}.${minor}.${patch}. Please upgrade to CouchDB 2.2.0 or later.`,
+    );
   }
 
   function designDocNeedsUpdate(newDesignDoc, oldDesignDoc) {
